huggingface/gradio_space/requirements.txt
gradio>=5.0.0
tensorflow==2.15.0
huggingface-hub>=0.20.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
